Keep arena::Range::inner field private.
